save an almost-certainly broken, but promising sketch.  This attempts to do everything I wanted except a package whitelist.  I initially implemented changes to bootstrap that removed site-packages, but decided that this was unnecessary, so ripped it out.  Tests are in progress, already causing some rewrites, but this is also a sketch for Jim to consider, if he gets around to it.

+++ bootstrap/bootstrap.py
@@ -20,19 +20,87 @@ use the -c option to specify an alternate configuration file.
 $Id$
 """
 
-import os, shutil, sys, tempfile, urllib2
+import os, re, shutil, sys, tempfile, textwrap, urllib, urllib2
 
-tmpeggs = tempfile.mkdtemp()
+# We have to manually parse our options rather than using one of the stdlib
+# tools because we want to pass the ones we don't recognize along to
+# zc.buildout.buildout.main.
 
-is_jython = sys.platform.startswith('java')
+configuration = {
+    '--ez_setup-source': 'http://peak.telecommunity.com/dist/ez_setup.py',
+    '--version': '',
+    '--download-base': None,
+    '--eggs': None}
+
+helpstring = __doc__ + textwrap.dedent('''
+    Options: 
+      --version=ZC_BUILDOUT_VERSION
+                Specify a version number of the zc.buildout to use
+      --ez_setup-source=URL_OR_FILE
+                Specify a URL or file location for the ez_setup file.
+                Defaults to
+                %(--ez_setup-source)s
+      --download-base=URL_OR_DIRECTORY
+                Specify a URL or directory for downloading setuptools and
+                zc.buildout.  Defaults to PyPI.
+      --eggs=DIRECTORY
+                Specify a directory for storing eggs.  Defaults to a temporary
+                directory that is deleted when the bootstrap script completes.
+
+    By using --ez_setup-source and --download-base to point to local resources,
+    you can keep bootstrap from going over the network.
+    ''')
+match_equals = re.compile(r'(%s)=(\S*)' % ('|'.join(configuration),)).match
+args = sys.argv[1:]
+if args == ['--help']:
+    print helpstring
+    sys.exit(0)
+
+# defaults
+tmpeggs = None
+
+while args:
+    val = args[0]
+    elif val in configuration:
+        del args[0]
+        if not args or args[0].startswith('-'):
+            print "ERROR: %s requires an argument."
+            print helpstring
+            sys.exit(1)
+        configuration[val] = args[0]
+    else:
+        match = match_equals(val)
+        if match and match.group(1) in configuration:
+            configuration[match.group(1)] = match.group(2)
+        else:
+            break
+    del args[0]
+
+for name in ('--ez_setup-source', '--download-base'):
+    val = configuration[name]
+    if val is not None and '://' not in val: # we're being lazy.
+        configuration[name] = 'file://%s' % (
+            urllib.pathname2url(os.path.abspath(os.path.expanduser(val))),)
+
+if not configuration['--eggs']:
+    configuration['--eggs'] = tmpeggs = tempfile.mkdtemp()
+else:
+    configuration['--eggs'] = os.path.abspath(
+        os.path.expanduser(configuration['--eggs']))
+
+if configuration['--version']:
+    configuration['--version'] = '==' + configuration['--version']
 
 try:
     import pkg_resources
 except ImportError:
     ez = {}
-    exec urllib2.urlopen('http://peak.telecommunity.com/dist/ez_setup.py'
-                         ).read() in ez
-    ez['use_setuptools'](to_dir=tmpeggs, download_delay=0)
+    exec urllib2.urlopen(configuration['--ez_setup-source']).read() in ez
+    setuptools_args = dict(to_dir=configuration['--eggs'], download_delay=0)
+    if configuration['--download-base']:
+        setuptools_args['download_base'] = (
+            configuration['--download-base'] + '/')
+    ez['use_setuptools'](**setuptools_args)
 
     import pkg_resources
 
@@ -45,40 +113,33 @@ if sys.platform == 'win32':
 else:
     def quote (c):
         return c
+cmd = [quote(sys.executable),
+       '-c',
+       quote('from setuptools.command.easy_install import main; main()'),
+       '-mqNxd',
+       quote(configuration['--eggs']),
+       'zc.buildout' + configuration['--version']]
+ws = pkg_resources.working_set
+env = dict(
+    os.environ,
+    PYTHONPATH=ws.find(pkg_resources.Requirement.parse('setuptools')).location)
 
-cmd = 'from setuptools.command.easy_install import main; main()'
-ws  = pkg_resources.working_set
-
-if len(sys.argv) > 2 and sys.argv[1] == '--version':
-    VERSION = '==%s' % sys.argv[2]
-    args = sys.argv[3:] + ['bootstrap']
-else:
-    VERSION = ''
-    args = sys.argv[1:] + ['bootstrap']
-
-if is_jython:
+try:
     import subprocess
-
-    assert subprocess.Popen([sys.executable] + ['-c', quote(cmd), '-mqNxd',
-           quote(tmpeggs), 'zc.buildout' + VERSION],
-           env=dict(os.environ,
-               PYTHONPATH=
-               ws.find(pkg_resources.Requirement.parse('setuptools')).location
-               ),
-           ).wait() == 0
-
+except ImportError:
+    exitcode = os.spawnle(*([os.P_WAIT, sys.executable] + cmd + [env]))
 else:
-    assert os.spawnle(
-        os.P_WAIT, sys.executable, quote (sys.executable),
-        '-c', quote (cmd), '-mqNxd', quote (tmpeggs), 'zc.buildout' + VERSION,
-        dict(os.environ,
-            PYTHONPATH=
-            ws.find(pkg_resources.Requirement.parse('setuptools')).location
-            ),
-        ) == 0
+    # Jython can use subprocess but not spawn.  We prefer it generally.
+    exitcode = subprocess.Popen(cmd, env=env).wait()
+if exitcode != 0:
+    # we shouldn't need an error message because a failure
+    # should have generated a visible traceback in the subprocess.
+    sys.exit(exitcode)
 
-ws.add_entry(tmpeggs)
-ws.require('zc.buildout' + VERSION)
+ws.add_entry(configuration['--eggs'])
+ws.require('zc.buildout' + configuration['--version'])
 import zc.buildout.buildout
+args.append('bootstrap')
 zc.buildout.buildout.main(args)
-shutil.rmtree(tmpeggs)
+if tmpeggs is not None:
+    shutil.rmtree(tmpeggs)
